IR receiver 8 channel coffin
The coffin was the centerpiece of my gallery. I designed it so the first three times a player shot the target, the coffin lid would shake and it played a 'monster groan'. On the fourth shot, the coffin lid opened, disco lights and music played, and a skull head bounced along with the music. Most people didn't pick up on the pattern and thought the lid opening was random.